Moved serializers to own folder
